MOBILE-77 Add Caching for NFT tracks upon connecting a wallet
