+++ android/features/music-player/src/main/java/io/newm/feature/musicplayer/viewmodel/MusicPlayerViewModel.kt
@@ -7,12 +7,15 @@ import io.newm.feature.musicplayer.repository.MusicRepository
 import io.newm.feature.musicplayer.service.MusicPlayer
 import io.newm.shared.public.models.NFTTrack
 import io.newm.shared.public.usecases.WalletNFTTracksUseCase
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 
@@ -25,17 +28,13 @@ class MusicPlayerViewModel(
     private val songIdFlow: MutableStateFlow<String?> = MutableStateFlow(songId)
 
     private val _state: StateFlow<MusicPlayerState> by lazy {
-        val songFlow = songIdFlow.flatMapLatest { songId ->
-            useCase.getAllNFTTracksFlow().map { songs ->
-                val songMap: Map<String, NFTTrack> = songs.associateBy { song -> song.name }
-                val songToPlay = songMap[songId]
-                songToPlay ?: songs.first()
-            }
 
+        val songFlow: Flow<NFTTrack?> = songIdFlow.filterNotNull().map { trackId ->
+            useCase.getNFTTrack(trackId)
         }
         combine(
             musicPlayer.playbackStatus,
-            songFlow
+            songFlow.mapNotNull { it }
         ) { playbackStatus, song ->
             songIdFlow.update { playbackStatus.track?.id }
