+++ buildSrc/src/main/java/SqlDelight.kt
@@ -1,5 +1,6 @@
 object SqlDelight {
     const val runtime = "com.squareup.sqldelight:runtime:${Versions.sqlDelight}"
+    const val coroutinesExtensions = "com.squareup.sqldelight:coroutines-extensions:${Versions.sqlDelight}"
     const val androidDriver = "com.squareup.sqldelight:android-driver:${Versions.sqlDelight}"
     const val nativeDriver = "com.squareup.sqldelight:native-driver:${Versions.sqlDelight}"
     const val gradlePlugin = "com.squareup.sqldelight:gradle-plugin:${Versions.sqlDelight}"
