--- shared/src/commonMain/kotlin/io.newm.shared/internal/entity/AuthData.kt
@@ -1,9 +1,0 @@
-package io.newm.shared.internal.entity
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AuthData(
-    val accessToken: String,
-    val refreshToken: String
-)
