+++ shared/src/commonMain/kotlin/io.newm.shared/internal/implementations/ConnectWalletUseCaseImpl.kt
@@ -1,20 +1,29 @@
 package io.newm.shared.internal.implementations
 
+import io.newm.shared.internal.repositories.CardanoWalletRepository
 import io.newm.shared.internal.repositories.ConnectWalletManager
 import io.newm.shared.public.usecases.ConnectWalletUseCase
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
 
 internal class ConnectWalletUseCaseImpl(
-    private val repository: ConnectWalletManager,
+    private val connectWalletManager: ConnectWalletManager,
+    private val cardanoWalletRepository: CardanoWalletRepository
 ) : ConnectWalletUseCase {
     override fun connect(xpub: String) {
-        repository.connect(xpub)
+        connectWalletManager.connect(xpub)
     }
 
     override fun disconnect() {
-        repository.disconnect()
+        connectWalletManager.disconnect()
+        cardanoWalletRepository.deleteAllNFTs()
     }
 
     override fun isConnected(): Boolean {
-        return repository.isConnected()
+        return connectWalletManager.isConnected()
+    }
+
+    override fun isConnectedFlow(): Flow<Boolean> = flow {
+        emit(connectWalletManager.isConnected())
     }
 }
