--- shared/src/commonMain/sqldelight/io.newm.shared.db.cache/NewmAuth.sq
@@ -1,20 +1,0 @@
-CREATE TABLE AuthEntity(
-    id INTEGER NOT NULL PRIMARY KEY,
-    access_token TEXT NOT NULL,
-    refresh_token TEXT NOT NULL
-);
-
-selectAll:
-SELECT *
-FROM AuthEntity;
-
-insertAuthData:
-INSERT OR REPLACE
-INTO AuthEntity(
-    id,
-    access_token,
-    refresh_token
-) VALUES (?, ?, ?);
-
-deleteAuthData:
-DELETE FROM AuthEntity;
